+++ lib/scope.dart
@@ -1,0 +1,7 @@
+export 'src/exceptions.dart'
+    show
+        MissingDependencyException,
+        CircularDependencyException,
+        DuplicateDependencyException;
+export 'src/scope.dart'
+    show Scope, ScopeKey, use, hasScopeKey, isWithinScope, isNullable;
