--- lib/zone_di2.dart
@@ -1,11 +1,0 @@
-export 'src/zone_di.dart'
-    show
-        Scope,
-        ScopeKey,
-        ValueFactory,
-        // provide,
-        // provideSingle,
-        // provideFactories,
-        use,
-        MissingDependencyException,
-        CircularDependencyException;
